back-end/requirements.txt
click==7.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Werkzeug==1.0.1
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
Flask==1.1.2
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
Flask-Restless==1.0.0b1
Flask-SQLAlchemy==2.4.0
pymysql
mysql-connector-python==8.0.18
mysqlclient==1.4.6
psycopg2==2.8.6
pyOpenSSL==19.1.0
python-dotenv==0.14.0
python-dateutil==2.8.0
python-mimeparse==1.6.0
PyYAML==5.2
SQLAlchemy==1.3.15
websocket-client==0.57.0
